requirements.txt
aiogram==3.7.0
motor==3.4.0
pymongo==4.7.2
openai>=1.52.0
httpx>=0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1